requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3

xgboost>=1.7
lightgbm>=4.0
catboost>=1.2
shap>=0.44

matplotlib>=3.7
seaborn>=0.12

pymatgen>=2023.9
openpyxl>=3.1

pytest>=7.4
pyyaml>=6.0